+++ src/Faker/Provider/uk_UA/Transliteration.php
@@ -1,0 +1,60 @@
+<?php
+
+namespace Faker\Provider\ua_UA;
+
+/**
+ * Cyrillic to latin transliteration util
+ */
+class Transliteration
+{
+    /**
+     * Translate cyrillic text to latin
+     * Useful for URL ang email generation
+     * @param  string $string
+     * @return string
+     */
+    public static function transliterate($string = '')
+    {
+        $replace=array(
+            "'"=>"",
+            "`"=>"",
+            "а"=>"a","А"=>"a",
+            "б"=>"b","Б"=>"b",
+            "в"=>"v","В"=>"v",
+            "г"=>"g","Г"=>"g",
+            "д"=>"d","Д"=>"d",
+            "е"=>"e","Е"=>"e",
+            "ж"=>"zh","Ж"=>"zh",
+            "з"=>"z","З"=>"z",
+            "и"=>"i","И"=>"i",
+            "й"=>"y","Й"=>"y",
+            "к"=>"k","К"=>"k",
+            "л"=>"l","Л"=>"l",
+            "м"=>"m","М"=>"m",
+            "н"=>"n","Н"=>"n",
+            "о"=>"o","О"=>"o",
+            "п"=>"p","П"=>"p",
+            "р"=>"r","Р"=>"r",
+            "с"=>"s","С"=>"s",
+            "т"=>"t","Т"=>"t",
+            "у"=>"u","У"=>"u",
+            "ф"=>"f","Ф"=>"f",
+            "х"=>"h","Х"=>"h",
+            "ц"=>"c","Ц"=>"c",
+            "ч"=>"ch","Ч"=>"ch",
+            "ш"=>"sh","Ш"=>"sh",
+            "щ"=>"sch","Щ"=>"sch",
+            "ъ"=>"","Ъ"=>"",
+            "ы"=>"y","Ы"=>"y",
+            "ь"=>"","Ь"=>"",
+            "э"=>"e","Э"=>"e",
+            "ю"=>"yu","Ю"=>"yu",
+            "я"=>"ya","Я"=>"ya",
+            "і"=>"i","І"=>"i",
+            "ї"=>"yi","Ї"=>"yi",
+            "є"=>"e","Є"=>"e"
+        );
+
+        return strtr($string, $replace);
+    }
+}
